requirements.txt
# ============================
# Medical Report Intelligence Engine
# ClinIQ - Dependencies
# ============================

# --- PDF & Document Handling ---
pdfplumber>=0.10.0
pytesseract>=0.3.10
reportlab>=4.0
python-docx>=1.0.0
Pillow>=10.0.0

# --- Backend API ---
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9

# --- LLM Inference ---
groq>=0.4.0
python-dotenv>=1.0.0

# --- RAG & Embeddings ---
chromadb>=0.4.22
sentence-transformers>=2.5.0

# --- PII / Privacy ---
presidio-analyzer>=2.2.0
presidio-anonymizer>=2.2.0
spacy>=3.7.0

# --- Data Validation ---
pydantic>=2.5.0

# --- Frontend ---
streamlit>=1.31.0
plotly>=5.18.0
requests>=2.31.0
